chore: rename to cssmodules
